back-end/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
supabase
Werkzeug
httpx
google-generativeai